refactor(share): move DataHash def from share/eds to share
