+++ share/eds/store.go
@@ -16,6 +16,8 @@ import (
 	bstore "github.com/ipfs/go-ipfs-blockstore"
 
 	"github.com/celestiaorg/rsmt2d"
+
+	"github.com/celestiaorg/celestia-node/share"
 )
 
 const (
@@ -140,7 +142,7 @@ func (s *Store) gc(ctx context.Context) {
 // The square is verified on the Exchange level, and Put only stores the square, trusting it.
 // The resulting file stores all the shares and NMT Merkle Proofs of the EDS.
 // Additionally, the file gets indexed s.t. store.Blockstore can access them.
-func (s *Store) Put(ctx context.Context, root DataHash, square *rsmt2d.ExtendedDataSquare) error {
+func (s *Store) Put(ctx context.Context, root share.DataHash, square *rsmt2d.ExtendedDataSquare) error {
 	// sanity check the root
 	if err := root.Validate(); err != nil {
 		return err
@@ -183,7 +185,7 @@ func (s *Store) Put(ctx context.Context, root DataHash, square *rsmt2d.ExtendedD
 // NMT Merkle proofs. Integrity of the store data is not verified.
 //
 // Caller must Close returned reader after reading.
-func (s *Store) GetCAR(ctx context.Context, root DataHash) (io.ReadCloser, error) {
+func (s *Store) GetCAR(ctx context.Context, root share.DataHash) (io.ReadCloser, error) {
 	key := root.String()
 	accessor, err := s.getAccessor(ctx, shard.KeyFromString(key))
 	if err != nil {
@@ -242,7 +244,7 @@ func (s *Store) getAccessor(ctx context.Context, key shard.Key) (*accessorWithBl
 
 // Remove removes EDS from Store by the given share.Root hash and cleans up all
 // the indexing.
-func (s *Store) Remove(ctx context.Context, root DataHash) error {
+func (s *Store) Remove(ctx context.Context, root share.DataHash) error {
 	key := root.String()
 
 	ch := make(chan dagstore.ShardResult, 1)
@@ -279,7 +281,7 @@ func (s *Store) Remove(ctx context.Context, root DataHash) error {
 //
 // It reads only one quadrant(1/4) of the EDS and verifies the integrity of the stored data by
 // recomputing it.
-func (s *Store) Get(ctx context.Context, root DataHash) (*rsmt2d.ExtendedDataSquare, error) {
+func (s *Store) Get(ctx context.Context, root share.DataHash) (*rsmt2d.ExtendedDataSquare, error) {
 	// sanity check the root
 	if err := root.Validate(); err != nil {
 		return nil, err
@@ -297,7 +299,7 @@ func (s *Store) Get(ctx context.Context, root DataHash) (*rsmt2d.ExtendedDataSqu
 }
 
 // Has checks if EDS exists by the given share.Root hash.
-func (s *Store) Has(ctx context.Context, root DataHash) (bool, error) {
+func (s *Store) Has(ctx context.Context, root share.DataHash) (bool, error) {
 	key := root.String()
 	info, err := s.dgstr.GetShardInfo(shard.KeyFromString(key))
 	if err == dagstore.ErrShardUnknown {
