+++ share/share.go
@@ -1,6 +1,8 @@
 package share
 
 import (
+	"fmt"
+
 	"go.opentelemetry.io/otel"
 
 	"github.com/celestiaorg/celestia-app/pkg/appconsts"
@@ -38,4 +40,19 @@ func ID(s Share) namespace.ID {
 // Data gets data from the share.
 func Data(s Share) []byte {
 	return s[NamespaceSize:]
+}
+
+// DataHash is a representation of the Root hash.
+type DataHash []byte
+
+func (dh DataHash) Validate() error {
+	if len(dh) != 32 {
+		return fmt.Errorf("invalid hash size, expected 32, got %d", len(dh))
+	}
+	return nil
+}
+
+func (dh DataHash) String() string {
+	c := dh
+	return fmt.Sprintf("%X", c)
 }
